This adds support for gomobile.
Internally gomobile calls "go build". By modifying the PATH env var before gomobile is called, we can get gomobile to call garble instead of go.

When the user runs "garble mobile", the garble binary is copied to a temp directory and named "go". The temp directory is prepended to the PATH env var.
When gomobile calls "go build" it will actually be calling garble.

When garble is run, it will always check if the string "garble" is in the path. If the string "garble" is found, the temp dir will
be removed from the PATH in order to avoid an infinite loop.

This is how the call flow looks like:
"garble mobile" -> "gomobile" -> "go" (garble renamed in a temp dir) -> "go" (garble calls the real go binary)

This implementation only supports the gomobile commands "bind" and "clean". The reason being that "gomobile build" will result in gomobile immediately trying to read the symbols
as soon as the binary has been built, but it can't since the symbols have already been stripped during compilation.

Code blocking "gomobile build":
https://github.com/golang/mobile/blob/35478a0c49da882188b186a3893d45be6ff74327/cmd/gomobile/build_androidapp.go#L87